Split code into SGVGridCell
