+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/SubjectGroupViewer/components/SGVGridCell/index.js
@@ -1,0 +1,1 @@
+export { default } from './SGVGridCell'
